# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AudienceRoom API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "operationId": "get_me_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/me/dashboard": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Dashboard",
        "operationId": "get_dashboard_users_me_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/session-participants": {
      "post": {
        "tags": [
          "session-participants"
        ],
        "summary": "Create Session Participant",
        "operationId": "create_session_participant_session_participants_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionParticipantCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionParticipantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "session-participants"
        ],
        "summary": "List Session Participants",
        "operationId": "list_session_participants_session_participants_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionParticipantResponse"
                  },
                  "title": "Response List Session Participants Session Participants Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session-participants/bulk": {
      "post": {
        "tags": [
          "session-participants"
        ],
        "summary": "Create Session Participants Bulk",
        "operationId": "create_session_participants_bulk_session_participants_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionParticipantBulkCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionParticipantResponse"
                  },
                  "type": "array",
                  "title": "Response Create Session Participants Bulk Session Participants Bulk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session-participants/{participant_id}": {
      "get": {
        "tags": [
          "session-participants"
        ],
        "summary": "Get Session Participant",
        "operationId": "get_session_participant_session_participants__participant_id__get",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionParticipantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiCharacterResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "strictness": {
            "type": "string",
            "title": "Strictness"
          },
          "personality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personality"
          },
          "voice_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Style"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "role",
          "strictness",
          "personality",
          "voice_style",
          "description",
          "is_active"
        ],
        "title": "AiCharacterResponse"
      },
...
      "DashboardResponse": {
        "properties": {
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions"
          },
          "completed_sessions": {
            "type": "integer",
            "title": "Completed Sessions"
          },
          "average_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score"
          },
          "recent_sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionListItem"
            },
            "type": "array",
            "title": "Recent Sessions"
          }
        },
        "type": "object",
        "required": [
          "total_sessions",
          "completed_sessions",
          "average_score",
          "recent_sessions"
        ],
        "title": "DashboardResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "overall_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Score"
          },
          "has_feedback": {
            "type": "boolean",
            "title": "Has Feedback"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "mode",
          "theme",
          "overall_score",
          "has_feedback",
          "started_at",
          "ended_at",
          "created_at"
        ],
        "title": "SessionListItem",
        "description": "Lightweight item for session history list."
      },
...
      "SessionParticipantBulkCreateRequest": {
        "properties": {
          "participants": {
            "items": {
              "$ref": "#/components/schemas/SessionParticipantCreateRequest"
            },
            "type": "array",
            "title": "Participants"
          }
        },
        "type": "object",
        "required": [
          "participants"
        ],
        "title": "SessionParticipantBulkCreateRequest"
      },
      "SessionParticipantCreateRequest": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "ai_character_id": {
            "type": "integer",
            "title": "Ai Character Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "seat_index": {
            "type": "integer",
            "title": "Seat Index"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "ai_character_id",
          "display_name",
          "role",
          "seat_index"
        ],
        "title": "SessionParticipantCreateRequest"
      },
      "SessionParticipantResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "ai_character_id": {
            "type": "integer",
            "title": "Ai Character Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "seat_index": {
            "type": "integer",
            "title": "Seat Index"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "ai_character": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiCharacterResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "ai_character_id",
          "display_name",
          "role",
          "seat_index",
          "is_active",
          "created_at"
        ],
        "title": "SessionParticipantResponse"
      },
      "UserCreateRequest": {
        "properties": {
          "firebase_uid": {
            "type": "string",
            "title": "Firebase Uid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "firebase_uid",
          "email",
          "display_name"
        ],
        "title": "UserCreateRequest"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "firebase_uid": {
            "type": "string",
            "title": "Firebase Uid"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "firebase_uid",
          "email",
          "display_name",
          "photo_url",
          "onboarding_completed",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}